# Transactional storage - Tests & RFC