--- a/sdk/physical/testing.go
+++ b/sdk/physical/testing.go
@@ -483,3 +483,257 @@ func ExerciseHABackend(t testing.TB, b HABackend, b2 HABackend) {
 	// Cleanup
 	lock2.Unlock()
 }
+
+func ExerciseTransactionalBackend(t testing.TB, b TransactionalBackend) {
+	t.Helper()
+
+	// Creating a transaction and committing or rolling it back without doing
+	// anything should succeed, regardless of type of transaction. Doing the
+	// same operation twice should fail as the transaction was already
+	// finished.
+	txn, err := b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin read-write transaction")
+	err = txn.Commit(context.Background())
+	require.NoError(t, err, "failed to commit transaction with no entries")
+	err = txn.Commit(context.Background())
+	require.Error(t, err, "expected double commit of transaction to fail")
+
+	txn, err = b.BeginReadOnlyTx(context.Background())
+	require.NoError(t, err, "failed to begin read-only transaction")
+	err = txn.Commit(context.Background())
+	require.NoError(t, err, "failed to commit read-only transaction with no entries")
+	err = txn.Commit(context.Background())
+	require.Error(t, err, "expected double commit of read-only transaction to fail")
+
+	txn, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second read-write transaction")
+	err = txn.Rollback(context.Background())
+	require.NoError(t, err, "failed to rollback transaction with no entries")
+	err = txn.Rollback(context.Background())
+	require.Error(t, err, "expected double rollback of transaction to fail")
+
+	txn, err = b.BeginReadOnlyTx(context.Background())
+	require.NoError(t, err, "failed to begin second read-only transaction")
+	err = txn.Rollback(context.Background())
+	require.NoError(t, err, "failed to rollback read-only transaction with no entries")
+	err = txn.Rollback(context.Background())
+	require.Error(t, err, "expected double rollback of read-only transaction to fail")
+
+	// This should also be true if we swap types (commit->rollback and
+	// visa-versa).
+	txn, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin read-write transaction")
+	err = txn.Commit(context.Background())
+	require.NoError(t, err, "failed to commit transaction with no entries")
+	err = txn.Rollback(context.Background())
+	require.Error(t, err, "expected subsequent rollback of transaction to fail")
+
+	txn, err = b.BeginReadOnlyTx(context.Background())
+	require.NoError(t, err, "failed to begin read-only transaction")
+	err = txn.Commit(context.Background())
+	require.NoError(t, err, "failed to commit read-only transaction with no entries")
+	err = txn.Rollback(context.Background())
+	require.Error(t, err, "expected subsequent rollback of read-only transaction to fail")
+
+	txn, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second read-write transaction")
+	err = txn.Rollback(context.Background())
+	require.NoError(t, err, "failed to rollback transaction with no entries")
+	err = txn.Commit(context.Background())
+	require.Error(t, err, "expected subsequent commit of transaction to fail")
+
+	txn, err = b.BeginReadOnlyTx(context.Background())
+	require.NoError(t, err, "failed to begin second read-only transaction")
+	err = txn.Rollback(context.Background())
+	require.NoError(t, err, "failed to rollback read-only transaction with no entries")
+	err = txn.Commit(context.Background())
+	require.Error(t, err, "expected subsequent commit of read-only transaction to fail")
+
+	// Empty transactions can be interwoven.
+	txn1, err := b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin first interwoven transaction")
+	txn2, err := b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second interwoven transaction")
+	err = txn2.Commit(context.Background())
+	require.NoError(t, err, "failed to commit second interwoven transaction")
+	err = txn1.Commit(context.Background())
+	require.NoError(t, err, "failed to commit second interwoven transaction")
+
+	// Writing to a read-only transaction should fail; committing this
+	// transaction should have no impact on storage.
+	entry := &Entry{Key: "foo", Value: []byte("test")}
+
+	rtx, err := b.BeginReadOnlyTx(context.Background())
+	require.NoError(t, err, "failed to create read-only transaction for writing")
+	err = rtx.Put(context.Background(), entry)
+	require.Error(t, err, "expected failure to put in read-only transaction")
+	err = rtx.Delete(context.Background(), "foo")
+	require.Error(t, err, "expected failure to delete in read-only transaction")
+
+	err = rtx.Commit(context.Background())
+	require.NoError(t, err, "failed to commit empty read-only transaction")
+
+	entries, err := b.List(context.Background(), "")
+	require.NoError(t, err, "failed to list storage entries")
+	require.Empty(t, entries, "expected nothing in storage")
+
+	// Creating the same entry in two transactions should conflict the
+	// second committed one, even though they have the same contents.
+
+	txn1, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin first conflicting transaction")
+	txn2, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second conflicting transaction")
+
+	err = txn1.Put(context.Background(), entry)
+	require.NoError(t, err, "unexpected failure writing to first transaction")
+	err = txn2.Put(context.Background(), entry)
+	require.NoError(t, err, "unexpected failure writing to second transaction")
+
+	err = txn1.Commit(context.Background())
+	require.NoError(t, err, "failed to commit first conflicting transaction")
+	err = txn2.Commit(context.Background())
+	require.Error(t, err, "expected failure to commit second conflicting transaction")
+
+	result, err := b.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry")
+	require.Equal(t, result.Value, []byte("test"))
+
+	bazEntry := &Entry{Key: "foo", Value: []byte("baz")}
+
+	txn1, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin first conflicting transaction (round 2)")
+	txn2, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second conflicting transaction (round 2)")
+
+	err = txn1.Put(context.Background(), bazEntry)
+	require.NoError(t, err, "unexpected failure writing to first transaction (round 2)")
+	err = txn2.Put(context.Background(), bazEntry)
+	require.NoError(t, err, "unexpected failure writing to second transaction (round 2)")
+
+	err = txn2.Commit(context.Background())
+	require.NoError(t, err, "failed to commit second conflicting transaction")
+	err = txn1.Commit(context.Background())
+	require.Error(t, err, "expected failure to commit first conflicting transaction")
+
+	result, err = b.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry")
+	require.Equal(t, result.Value, []byte("baz"))
+
+	// Creating different entries in two transactions should be fine.
+
+	txn1, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin first parallel transaction")
+	txn2, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second parallel transaction")
+
+	barEntry := &Entry{Key: "bar", Value: []byte("baz")}
+	err = txn1.Put(context.Background(), entry)
+	require.NoError(t, err, "unexpected failure writing to first transaction")
+	err = txn2.Put(context.Background(), barEntry)
+	require.NoError(t, err, "unexpected failure writing to second transaction")
+
+	err = txn1.Commit(context.Background())
+	require.NoError(t, err, "failed to commit first parallel transaction")
+	err = txn2.Commit(context.Background())
+	require.NoError(t, err, "failed to commit second parallel transaction")
+
+	result, err = b.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry")
+	require.Equal(t, result.Value, []byte("test"))
+
+	result, err = b.Get(context.Background(), "bar")
+	require.NoError(t, err, "failed to read storage entry")
+	require.Equal(t, result.Value, []byte("baz"))
+
+	// Getting an item and writing to the same item in different transactions
+	// should fail one of the two.
+
+	txn1, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin first parallel transaction")
+	txn2, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second parallel transaction")
+
+	result, err = txn1.Get(context.Background(), "bar")
+	require.NoError(t, err, "failed to read storage entry in first transaction")
+	require.Equal(t, result.Value, []byte("baz"))
+	err = txn1.Put(context.Background(), bazEntry)
+	require.NoError(t, err, "failed to put storage entry in first transaction")
+
+	result, err = txn2.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry in second transaction")
+	require.Equal(t, result.Value, []byte("test"))
+	barTestEntry := &Entry{Key: "bar", Value: []byte("test")}
+	err = txn2.Put(context.Background(), barTestEntry)
+	require.NoError(t, err, "failed to put storage entry in second transaction")
+
+	err = txn1.Commit(context.Background())
+	require.NoError(t, err, "failed to commit first conflicting read transaction")
+	err = txn2.Commit(context.Background())
+	require.Error(t, err, "expected failure to commit second conflicting read transaction")
+
+	result, err = b.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry")
+	require.Equal(t, result.Value, []byte("baz"))
+
+	result, err = b.Get(context.Background(), "bar")
+	require.NoError(t, err, "failed to read storage entry")
+	require.Equal(t, result.Value, []byte("baz"))
+
+	// Try again, with delete this time, committing in a different order.
+
+	txn1, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin first parallel transaction")
+	txn2, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin second parallel transaction")
+
+	result, err = txn1.Get(context.Background(), "bar")
+	require.NoError(t, err, "failed to read storage entry in first transaction")
+	require.Equal(t, result.Value, []byte("baz"))
+	err = txn1.Delete(context.Background(), "foo")
+	require.NoError(t, err, "failed to delete storage entry in first transaction")
+
+	result, err = txn2.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry in second transaction")
+	require.Equal(t, result.Value, []byte("baz"))
+	err = txn2.Delete(context.Background(), "bar")
+	require.NoError(t, err, "failed to delete storage entry in second transaction")
+
+	err = txn2.Commit(context.Background())
+	require.NoError(t, err, "failed to commit second conflicting read transaction")
+	err = txn1.Commit(context.Background())
+	require.Error(t, err, "expected failure to commit first conflicting read transaction")
+
+	result, err = b.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry")
+	require.Equal(t, result.Value, []byte("baz"))
+
+	entries, err = b.List(context.Background(), "")
+	require.NoError(t, err, "failed to list storage entries")
+	require.Equal(t, entries, []string{"foo"}, "expected only a single entry in storage")
+
+	// Reading entries that don't exist shouldn't cause issues.
+	txn, err = b.BeginTx(context.Background())
+	require.NoError(t, err, "failed to begin empty-read transaction")
+
+	result, err = txn.Get(context.Background(), "bar")
+	require.NoError(t, err, "failed to read storage entry in first transaction")
+	if result != nil {
+		require.Equal(t, len(result.Value), 0, "expected empty storage entry `bar`")
+	}
+
+	result, err = txn.Get(context.Background(), "foo")
+	require.NoError(t, err, "failed to read storage entry in second transaction")
+	require.Equal(t, result.Value, []byte("baz"))
+
+	err = txn.Delete(context.Background(), "foo")
+	require.NoError(t, err, "failed to delete entry in transaction")
+
+	err = txn.Commit(context.Background())
+	require.NoError(t, err, "failed to commit deletion transaction")
+
+	// Ensure we left it as we found it.
+	entries, err = b.List(context.Background(), "")
+	require.NoError(t, err, "failed to list storage entries")
+	require.Empty(t, entries, "expected nothing in storage")
+}