This adds support for interactive transactions across the backend, physical, and plugin storage levels. This will ultimately allow for greater assurance in secrets management operations built on top of OpenBao.

See the RFC in #296 for more detailed information about the design of this change.

Resolves: #296

----

TODO:

 - [x] Resolve unresolved questions. 
 - [x] Benchmark subsystem performance when using transactions.
 - [x] Implement support for GRPC connected plugins.
 - [x] Expand randomized testing to `logical.Storage`, finish adding remaining layers to the cross-test. 
 - [x] Obey error handling discussed in RFC. 
 - [x] Add transaction concurrency limiting to InMem as well. 
 - [x] Expand `physical`'s `testing.go` to support basic testing of transactions. 
 - [x] ~Add support for disabling transactional storage from the configuration.~
 - [x] ~Update inmem-ha, `sdk/physical/latency.go`, and `sdk/physical/errors.go` to be transactional.~
 - [x] Think about separate raft limit size of transactions. 
 - [x] Rebase after GA and split PRs land.
    - [x] Split off removal of one-shot transactional interface into separate PR. -- #422 
    - [x] Split seal unwrapper removal into separate PR.  -- #423
    - [x] Split off core transaction definitions -- #433
    - [x] Split off each implementation -- #436, #437, #438, #439, #441, #442, and #443.
    - [x] Split off second level implementations (`BarrierView` and `AES-GCM Barrier View`) -- #440, #495
    - [x] Split off GRPC implementation -- #497
    - [x] Split off PKI root CA creation updates -- #498
    - [x] Add RFC text in this PR, merge.

<details>

<summary> Benchmarks </summary>

Without transactions (b56dc9c110b7b4f9170a1148d9febcc9a35cef38):

```
$ vault-benchmark run -config=docs/examples/pki/stored/pki-ed25519.hcl 
2024-07-22T12:28:57.564-0400 [INFO]  vault-benchmark: setting up targets
2024-07-22T12:29:01.596-0400 [INFO]  vault-benchmark: starting benchmarks: duration=30s
2024-07-22T12:29:31.604-0400 [INFO]  vault-benchmark: benchmark complete
Target: http://127.0.0.1:8200
op         count   rate         throughput   mean        95th%       99th%       successRatio
pki_issue  216959  7231.820942  7230.238976  1.363455ms  3.034576ms  5.149475ms  100.00%
```


With transactions (unused):

```
$ vault-benchmark run -config=docs/examples/pki/stored/pki-ed25519.hcl 
2024-07-22T12:27:22.730-0400 [INFO]  vault-benchmark: setting up targets
2024-07-22T12:27:26.775-0400 [INFO]  vault-benchmark: starting benchmarks: duration=30s
2024-07-22T12:27:56.779-0400 [INFO]  vault-benchmark: benchmark complete
Target: http://127.0.0.1:8200
op         count   rate         throughput   mean        95th%       99th%       successRatio
pki_issue  218338  7277.948872  7277.202214  1.356165ms  3.004142ms  4.872949ms  100.00%
```

With transactions (used for PKI issuance):

```
$ vault-benchmark run -config=docs/examples/pki/stored/pki-ed25519.hcl 
2024-07-22T12:33:15.869-0400 [INFO]  vault-benchmark: setting up targets
2024-07-22T12:33:19.896-0400 [INFO]  vault-benchmark: starting benchmarks: duration=30s
2024-07-22T12:33:49.898-0400 [INFO]  vault-benchmark: benchmark complete
Target: http://127.0.0.1:8200
op         count   rate         throughput   mean        95th%       99th%       successRatio
pki_issue  224037  7467.877003  7467.576805  1.318337ms  3.120374ms  5.828247ms  100.00%
```

<details>

<summary> Diff for PKI issuance transactions </summary>

```diff
$ git diff
diff --git a/builtin/logical/pki/path_issue_sign.go b/builtin/logical/pki/path_issue_sign.go
index 8c1a01451..7e554d129 100644
--- a/builtin/logical/pki/path_issue_sign.go
+++ b/builtin/logical/pki/path_issue_sign.go
@@ -388,6 +388,18 @@ func (b *backend) pathIssueSignCert(ctx context.Context, req *logical.Request, d
                return nil, logical.ErrReadOnly
        }
 
+       // If we have a transactional storage backend, let's use it.
+       if txnStorage, ok := req.Storage.(logical.TransactionalStorage); ok {
+               b.Logger().Debug("using transaction for root CA generation")
+               txn, err := txnStorage.BeginTx(ctx)
+               if err != nil {
+                       return nil, err
+               }
+
+               defer txn.Rollback(ctx)
+               req.Storage = txn
+       }
+
        // We prefer the issuer from the role in two cases:
        //
        // 1. On the legacy sign-verbatim paths, as we always provision an issuer
@@ -561,6 +573,14 @@ func (b *backend) pathIssueSignCert(ctx context.Context, req *logical.Request, d
                }
        }
 
+       // Finally, commit our transaction if we have one!
+       if txn, ok := req.Storage.(logical.Transaction); ok {
+               if err := txn.Commit(ctx); err != nil {
+                       return nil, err
+               }
+               resp.AddWarning("cert issued with transactions")
+       }
+
        resp = addWarnings(resp, warnings)
 
        return resp, nil
```

</details>

</details>