@cipherboy 2024-10-05
@DanGhita @JanMa I'm inclined to land this with just the tests and the RFCs now and then rework both as necessary, when we're landing PostgreSQL transactions.

However, I still think PostgreSQL transactions should be landed before we do 2.1.0 release.

Thoughts?